Make token getters not return null, but empty optional (#14)
* Make token getters not return null, but empty optional

I feel that returning null instead of an empty optional is counter intuitive. 

At it makes it necessary to actually do two checks:
1. Check if the optional is null
2. Check if the optional is empty

* Forgot to remove return statement

Sorry, should have tested

* Expect empty Optional instead of null

This in relation to my first pull request

* Removed check for preferred username match

It can not be generally assumed that the login name is identical to the preferred username. On a newly created production user, the preferred username is the first name + four digits. Where as I log in with the email address. So This check is not valid in my opinion.